+++ ios/Classes/Banner/BannerAdView.swift
@@ -1,0 +1,60 @@
+import Flutter
+import GoogleMobileAds
+import UIKit
+
+class BannerAdViewFactory: NSObject, FlutterPlatformViewFactory {
+    private var messenger: FlutterBinaryMessenger
+
+    init(messenger: FlutterBinaryMessenger) {
+        self.messenger = messenger
+        super.init()
+    }
+
+    func create(
+        withFrame frame: CGRect,
+        viewIdentifier viewId: Int64,
+        arguments args: Any?
+    ) -> FlutterPlatformView {
+        return BannerAdView(
+            frame: frame,
+            viewIdentifier: viewId,
+            arguments: args,
+            binaryMessenger: messenger)
+    }
+}
+
+class BannerAdView: NSObject, FlutterPlatformView {
+    private var _view: UIView
+    private var data: Any?
+    private var controller: BannerAdController
+
+    private var adSize: AdSize
+
+    init(
+        frame: CGRect,
+        viewIdentifier viewId: Int64,
+        arguments args: Any?,
+        binaryMessenger messenger: FlutterBinaryMessenger?
+    ) {
+        _view = UIView()
+        data = args
+        controller = BannerAdControllerManager.getController(data["controllerId"] as String)
+        super.init()
+        // iOS views can be created here
+        createNativeView(view: _view)
+    }
+
+    func view() -> UIView {
+        return _view
+    }
+
+    func createNativeView(view _view: UIView) {
+        _view.backgroundColor = UIColor.blue
+        let nativeLabel = UILabel()
+        nativeLabel.text = "Native text from iOS"
+        nativeLabel.textColor = UIColor.white
+        nativeLabel.textAlignment = .center
+        nativeLabel.frame = CGRect(x: 0, y: 0, width: 180, height: 48.0)
+        _view.addSubview(nativeLabel)
+    }
+}