+++ lib/native_admob_flutter.dart
@@ -1,5 +1,5 @@
 export 'src/mobile_ads.dart';
-export 'src/utils.dart' show AdError;
+export 'src/utils.dart' show AdError, LoadShowAd, UniqueKeyMixin;
 
 export 'src/native/native_ad_widget.dart';
 export 'src/native/controller/controller.dart';